Renamed PixiPusher and Color
